spice: extract certified accumulator into its own module
